% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state ()
this function that plots points where accidents for given state.num
in other words, the fars_map_state function takes a state ID number and a year, and maps that state's fatalities with a dot at the fatality location.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is state ID number (one interger from unique(data$STATE) if invalid state number used, the function will stop.}

\item{year}{is a single year as the argument for make_filename() function.}
}
\description{
internal functions needed are make_filename() and fars_read()
external packages needed are:  dplyr,maps, graphics. functiion filter() from dplyr, map() from maps and points () from graphics are used.
}
\examples{
state_num <- 49 # Utah
yr <- 2014

}
