% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{fars_read ()is a function that read in data from a csv file.
FARS stand for "Fatality Analysis Reporting System" and is a nationwide census, providing the American public yearly data,
regarding fatal injuries suffered in motor vehicle traffic crashes}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is for the file to be read. if this file name not exist, function will stop and print "does not exist".}
}
\value{
This function returns the data to be read
}
\description{
fars_read ()is a function that read in data from a csv file.
FARS stand for "Fatality Analysis Reporting System" and is a nationwide census, providing the American public yearly data,
regarding fatal injuries suffered in motor vehicle traffic crashes
}
\examples{
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
fars_read("accident_2015.csv.bz2")

}
