% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years is function that read in multiple files by years.
internal functions need for fars_read_years are make_filename () in this code.
external package needed is dplyr. Two functions of it are selection(),mutate().}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is a vector of years as the argument for the fars_read_years() function.
if a year not exist, warning message will print "invalid year".}
}
\value{
this function returns data files with file names defined
}
\description{
fars_read_years is function that read in multiple files by years.
internal functions need for fars_read_years are make_filename () in this code.
external package needed is dplyr. Two functions of it are selection(),mutate().
}
\examples{
fars_read_years(years = c(2013, 2015)
fars_read_years(2013:2015)

}
