% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years()
this function summarizes yearly accidents data, by month and produces a summary of the simple counts of fatalities by month and year
internal functions needed is fars_read_years()
external package needed is dplyr and tidyr. Thre functions of dplyr are used :sbind_rows(),group_by(),summarize().
function spread() from tidyr is used as well.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is a vector of years as the argument.
The fars_summarize_years() function take the same argument as the fars_read_years().}
}
\description{
fars_summarize_years()
this function summarizes yearly accidents data, by month and produces a summary of the simple counts of fatalities by month and year
internal functions needed is fars_read_years()
external package needed is dplyr and tidyr. Thre functions of dplyr are used :sbind_rows(),group_by(),summarize().
function spread() from tidyr is used as well.
}
\examples{
fars_summarize_years(years = c(2013, 2015)
fars_summarize_years(2013:2015)

}
