% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{make_filename is function that creates a filename based on a single year as input.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{is a single year in the data.}
}
\value{
this function returns a data file with file name defined
}
\description{
make_filename is function that creates a filename based on a single year as input.
}
\examples{
make_filename(2003)
make_filename(2004)

}
